Fixed skills

+++ TheRunes/src/main/java/com/qualityplus/runes/api/provider/LevelProvider.java
@@ -3,7 +3,7 @@ package com.qualityplus.runes.api.provider;
 import org.bukkit.entity.Player;
 
 public interface LevelProvider {
-    int getLevel(Player player);
+    double getLevel(Player player);
 
     void addXp(Player player, double xp);
 }
