+++ TheSkills/src/main/java/com/qualityplus/skills/base/perk/perks/BonusAttackSpeedPerk.java
@@ -28,9 +28,9 @@ public final class BonusAttackSpeedPerk extends Perk {
      */
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         List<IPlaceholder> placeholders = PlaceholderBuilder.create()
-                .with(new Placeholder("level_number", level), new Placeholder("level_roman", NumberUtil.toRoman(level)))
+                .with(new Placeholder("level_number", level), new Placeholder("level_roman", NumberUtil.toRoman((int)level)))
                 .with(new Placeholder("percent", chancePerLevel * level))
                 .get();
         return StringUtils.processMulti(description, placeholders);
