+++ TheSkills/src/main/java/com/qualityplus/skills/base/perk/perks/LightningPunchPerk.java
@@ -48,7 +48,7 @@ public final class LightningPunchPerk extends Perk {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         return StringUtils.processMulti(super.getFormattedDescription(level), PlaceholderBuilder.create(new Placeholder("damage", damage)).get());
     }
 }
