+++ TheSkills/src/main/java/com/qualityplus/skills/base/perk/perks/SpidermanPerk.java
@@ -74,7 +74,7 @@ public final class SpidermanPerk extends Perk {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         return StringUtils.processMulti(super.getFormattedDescription(level), PlaceholderBuilder.create(new Placeholder("duration", secondsDuration)).get());
     }
 }
