+++ TheSkills/src/main/java/com/qualityplus/skills/base/perk/perks/common/AbstractFortunePerk.java
@@ -31,14 +31,14 @@ public abstract class AbstractFortunePerk extends Perk {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         return StringUtils.processMulti(super.getFormattedDescription(level), PlaceholderBuilder.create(
                 new Placeholder("multiplier", getMultiplier(level)),
                 new Placeholder("chance", getChance(level)))
                 .get());
     }
 
-    protected int getMultiplier(int level) {
+    protected int getMultiplier(double level) {
         double chance = chancePerLevel * level;
 
         double add = 2;
@@ -49,7 +49,7 @@ public abstract class AbstractFortunePerk extends Perk {
         return (int) ((chance / 100) + add);
     }
 
-    protected double getChance(int level) {
+    protected double getChance(double level) {
         double chance = chancePerLevel * level;
 
         chance -= ((getMultiplier(level) - 2) * 100);
