+++ TheSkills/src/main/java/com/qualityplus/skills/base/perk/perks/common/AbstractPotionPerk.java
@@ -29,9 +29,9 @@ public abstract class AbstractPotionPerk extends Perk {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         return StringUtils.processMulti(super.getFormattedDescription(level), PlaceholderBuilder.create(
-                new Placeholder("duration", getDurationSeconds(level)),
+                new Placeholder("duration", getDurationSeconds((int)level)),
                 new Placeholder("potion_level_roman", NumberUtil.toRoman(getLevel())),
                 new Placeholder("potion_level_number", getLevel())
         ).get());
@@ -41,8 +41,8 @@ public abstract class AbstractPotionPerk extends Perk {
         return Math.max(0, level - 1);
     }
 
-    protected int getDurationTicks(int level) {
-        return baseSecondsDuration + (secondsDurationPerLevel * level) * 20;
+    protected int getDurationTicks(double level) {
+        return (int) (baseSecondsDuration + (secondsDurationPerLevel * level) * 20);
     }
 
     private int getDurationSeconds(int level) {
