+++ TheSkills/src/main/java/com/qualityplus/skills/base/stat/Stat.java
@@ -29,14 +29,14 @@ public abstract class Stat extends CommonObject implements ListenerRegistrable {
         extraListeners.add(listener);
     }
 
-    public int getStat(Player player, String id) {
+    public double getStat(Player player, String id) {
         return TheSkills.getApi().getSkillsService().getData(player.getUniqueId())
                 .map(UserData::getSkills)
                 .map(userPerks -> userPerks.getLevel(id))
-                .orElse(1);
+                .orElse(1D);
     }
 
-    public int getStat(Player player) {
+    public double getStat(Player player) {
         return getStat(player, id);
     }
 }
