+++ TheSkills/src/main/java/com/qualityplus/skills/base/stat/stats/CritChanceStat.java
@@ -72,10 +72,10 @@ public final class CritChanceStat extends Stat {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         List<IPlaceholder> placeholders = PlaceholderBuilder.create()
                 .with(new Placeholder("level_number", level),
-                        new Placeholder("level_roman", NumberUtil.toRoman(level)),
+                        new Placeholder("level_roman", NumberUtil.toRoman((int) level)),
                         new Placeholder("chance", chancePerLevel * level)
                 ).get();
         return StringUtils.processMulti(description, placeholders);
