+++ TheSkills/src/main/java/com/qualityplus/skills/base/stat/stats/IntelligenceStat.java
@@ -23,11 +23,11 @@ public final class IntelligenceStat extends Stat {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         List<IPlaceholder> placeholders = PlaceholderBuilder.create()
                 .with(new Placeholder("intelligence", level),
                       new Placeholder("level_number", level),
-                      new Placeholder("level_roman", NumberUtil.toRoman(level))
+                      new Placeholder("level_roman", NumberUtil.toRoman((int)level))
                 ).get();
         return StringUtils.processMulti(description, placeholders);
     }
