+++ TheSkills/src/main/java/com/qualityplus/skills/base/stat/stats/PetLuckStat.java
@@ -23,10 +23,10 @@ public final class PetLuckStat extends Stat {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         List<IPlaceholder> placeholders = PlaceholderBuilder.create()
                 .with(new Placeholder("level_number", level),
-                      new Placeholder("level_roman", NumberUtil.toRoman(level)),
+                      new Placeholder("level_roman", NumberUtil.toRoman((int)level)),
                       new Placeholder("chance", 0)
                 ).get();
         return StringUtils.processMulti(description, placeholders);
