+++ TheSkills/src/main/java/com/qualityplus/skills/base/stat/stats/StrengthStat.java
@@ -21,10 +21,10 @@ public final class StrengthStat extends Stat {
     }
 
     @Override
-    public List<String> getFormattedDescription(int level) {
+    public List<String> getFormattedDescription(double level) {
         List<IPlaceholder> placeholders = PlaceholderBuilder.create()
                 .with(new Placeholder("level_number", level),
-                      new Placeholder("level_roman", NumberUtil.toRoman(level)),
+                      new Placeholder("level_roman", NumberUtil.toRoman((int)level)),
                       new Placeholder("chance", 0),
                       new Placeholder("damage_percentage", 0)
