+++ TheSkills/src/main/java/com/qualityplus/skills/persistance/data/user/UserSkills.java
@@ -19,29 +19,29 @@ import java.util.Map;
 
 @Getter
 @Setter
-public final class UserSkills extends Document implements Levellable<String, Integer>, Progressable<String, Double>,
+public final class UserSkills extends Document implements Levellable<String, Double>, Progressable<String, Double>,
         LevellableArmorData<String> {
     @Exclude
     private Map<String, Integer> fromArmor = new HashMap<>();
-    private Map<String, Integer> level = new HashMap<>();
+    private Map<String, Double> level = new HashMap<>();
     private Map<String, Double> xp = new HashMap<>();
 
     public void fillIfEmpty() {
-        Skills.values().stream().map(Skill::getId).forEach(skill -> level.putIfAbsent(skill, 0));
+        Skills.values().stream().map(Skill::getId).forEach(skill -> level.putIfAbsent(skill, 0D));
         Skills.values().stream().map(Skill::getId).forEach(skill -> xp.putIfAbsent(skill, 0D));
         Skills.values().stream().map(Skill::getId).forEach(skill -> fromArmor.putIfAbsent(skill, 0));
 
         //Perks
-        Perks.values().stream().map(Perk::getId).forEach(perk -> level.putIfAbsent(perk, 0));
+        Perks.values().stream().map(Perk::getId).forEach(perk -> level.putIfAbsent(perk, 0D));
         Perks.values().stream().map(Perk::getId).forEach(perk -> fromArmor.putIfAbsent(perk, 0));
         //Stats
-        Stats.values().stream().map(Stat::getId).forEach(perk -> level.putIfAbsent(perk, 0));
+        Stats.values().stream().map(Stat::getId).forEach(perk -> level.putIfAbsent(perk, 0D));
         Stats.values().stream().map(Stat::getId).forEach(perk -> fromArmor.putIfAbsent(perk, 0));
     }
 
     @Override
-    public Integer getDefault() {
-        return 0;
+    public Double getDefault() {
+        return 0D;
     }
 
     @Override
